Created prtgDevices.js
Moved this function to it’s own .js file
